+++ Editor/Window/SWBuildReportViewerWindow.cs
@@ -128,7 +128,7 @@ namespace SW.EditorTools.Window
         /// <summary>
         /// Build Report Viewer 창을 엽니다.
         /// </summary>
-        [MenuItem("SWTools/Debug/Build Report Viewer")]
+        [MenuItem("SWTools/Debug/Build/Build Report Viewer")]
         public static void ShowWindow()
         {
             SWBuildReportViewerWindow window = GetWindow<SWBuildReportViewerWindow>();