+++ Editor/Window/SWPlayerPrefsViewerWindow.cs
@@ -83,7 +83,7 @@ namespace SW.EditorTools.Window
         /// <summary>
         /// PlayerPrefs Viewer 창을 엽니다.
         /// </summary>
-        [MenuItem("SWTools/Debug/PlayerPrefs Viewer")]
+        [MenuItem("SWTools/Debug/PlayerPrefs/PlayerPrefs Viewer")]
         public static void ShowWindow()
         {
             SWPlayerPrefsViewerWindow window = GetWindow<SWPlayerPrefsViewerWindow>();