+++ Editor/Window/SWTMPFontAssetManagerWindow.cs
@@ -182,7 +182,7 @@ namespace SW.EditorTools.Window
         /// <summary>
         /// TMP Font Asset Manager 창을 엽니다.
         /// </summary>
-        [MenuItem("SWTools/Utils/TMP Font Asset Manager")]
+        [MenuItem("SWTools/Utils/Asset/TMP Font Asset Manager")]
         public static void ShowWindow()
         {
             var window = GetWindow<SWTMPFontAssetManagerWindow>();